requirements.txt
groq>=0.4.1
python-dotenv>=1.0.0
datasets>=2.14.0
tqdm>=4.65.0
